refactor: split single state file into one per category
A bit overkill but it looks cleaner
