+++ src/common/state/totalSurplusState/index.ts
@@ -1,0 +1,2 @@
+export { TotalSurplusUpdater } from './updaters'
+export { useTriggerTotalSurplusUpdateCallback, useTotalSurplus } from './hooks'
