+++ src/common/state/totalSurplusState/types.ts
@@ -1,0 +1,10 @@
+import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
+
+import { Nullish } from 'types'
+
+export type TotalSurplusState = {
+  surplusAmount: Nullish<CurrencyAmount<Currency>>
+  isLoading: boolean
+  error: string
+  refetch: (() => void) | null
+}
